+++ example/lib/tabs/yt_web.dart
@@ -39,6 +39,7 @@ class YTWeb extends StatelessWidget {
             ),
             playerUIVisibilityOptions: PlayerUIVisibilityOptions().copyWith(
               useSafeAreaForBottomControls: true,
+              showPlaybackSpeedButton: true,
             ),
             customPlayerWidgets: CustomPlayerWidgets().copyWith(
               thumbnailFit: BoxFit.fitWidth,