+++ lib/src/video_player_initializer/network_initializer.dart
@@ -46,6 +46,8 @@ class NetworkInitializer implements IVideoPlayerInitializerStrategy {
         globalController: globalController,
         initialPosition: options.videoSourceConfiguration.initialPosition,
         initialVolume: options.videoSourceConfiguration.initialVolume,
+        initialPlaybackSpeed:
+            options.videoSourceConfiguration.initialPlaybackSpeed,
         callbacks: callbacks,
         type: options.videoSourceConfiguration.videoSourceType,
         globalKeyPlayer: options.globalKeyPlayer,